use Annotator in handler, cleanup dead code
