+++ handler/handler_test.go
@@ -27,7 +27,7 @@ func TestAnnotate(t *testing.T) {
 		{"1.4.128.0", "625600", `{"Geo":{"region":"ME","city":"Not A Real City","postal_code":"10583","latitude":42.1,"longitude":-73.1},"ASN":{}}`},
 		{"This will be an error.", "1000", "Invalid request"},
 	}
-	handler.CurrentGeoDataset = &parser.GeoDataset{
+	handler.CurrentAnnotator = &parser.GeoDataset{
 		IP4Nodes: []parser.IPNode{
 			{
 				IPAddressLow:  net.IPv4(0, 0, 0, 0),
@@ -183,7 +183,7 @@ func TestBatchAnnotate(t *testing.T) {
 			alt: `{"127.0.0.1ov94o0":{"Geo":{"region":"ME","city":"Not A Real City","postal_code":"10583","latitude":0,"longitude":0},"ASN":{}},"2620:0:1003:1008:5179:57e3:3c75:1886ov94o0":{"Geo":{"region":"ME","city":"Not A Real City","postal_code":"10583","latitude":0,"longitude":0},"ASN":{}}}`,
 		},
 	}
-	handler.CurrentGeoDataset = &parser.GeoDataset{
+	handler.CurrentAnnotator = &parser.GeoDataset{
 		IP4Nodes: []parser.IPNode{
 			{
 				IPAddressLow:  net.IPv4(0, 0, 0, 0),
@@ -231,7 +231,7 @@ func TestGetMetadataForSingleIP(t *testing.T) {
 				ASN: &common.IPASNData{}},
 		},
 	}
-	handler.CurrentGeoDataset = &parser.GeoDataset{
+	handler.CurrentAnnotator = &parser.GeoDataset{
 		IP4Nodes: []parser.IPNode{
 			{
 				IPAddressLow:  net.IPv4(0, 0, 0, 0),
@@ -259,46 +259,5 @@ func TestGetMetadataForSingleIP(t *testing.T) {
 		if !reflect.DeepEqual(res, test.res) {
 			t.Errorf("Expected %v, got %v", test.res, res)
 		}
-	}
-}
-
-func TestConvertIPNodeToGeoData(t *testing.T) {
-	tests := []struct {
-		node parser.IPNode
-		locs []parser.LocationNode
-		res  *common.GeoData
-	}{
-		{
-			node: parser.IPNode{LocationIndex: 0, PostalCode: "10583"},
-			locs: []parser.LocationNode{{CityName: "Not A Real City", RegionCode: "ME"}},
-			res: &common.GeoData{
-				Geo: &common.GeolocationIP{City: "Not A Real City", Postal_code: "10583", Region: "ME"},
-				ASN: &common.IPASNData{}},
-		},
-		{
-			node: parser.IPNode{LocationIndex: -1, PostalCode: "10583"},
-			locs: nil,
-			res: &common.GeoData{
-				Geo: &common.GeolocationIP{Postal_code: "10583"},
-				ASN: &common.IPASNData{}},
-		},
-	}
-	for _, test := range tests {
-		res := handler.ConvertIPNodeToGeoData(test.node, test.locs)
-		if !reflect.DeepEqual(res, test.res) {
-			t.Errorf("Expected %v, got %v", test.res, res)
-		}
-	}
-}
-
-func TestExtractDateFromFilename(t *testing.T) {
-	date, err := handler.ExtractDateFromFilename("Maxmind/2017/05/08/20170508T080000Z-GeoLiteCity.dat.gz")
-	if date.Year() != 2017 || date.Month() != 5 || date.Day() != 8 || err != nil {
-		t.Errorf("Did not extract data correctly. Expected %d, got %v, %+v.", 20170508, date, err)
-	}
-
-	date2, err := handler.ExtractDateFromFilename("Maxmind/2017/10/05/20171005T033334Z-GeoLite2-City-CSV.zip")
-	if date2.Year() != 2017 || date2.Month() != 10 || date2.Day() != 5 || err != nil {
-		t.Errorf("Did not extract data correctly. Expected %d, got %v, %+v.", 20171005, date2, err)
 	}
 }
